+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+2020-10-15.01
+-------------
+
+* Items: add `item_analysis_title`, `item_analysis_content`
+
 2020-10-14.01
 -------------